+++ internal/app/album/delivery/http/album_handler.go
@@ -141,7 +141,7 @@ func (h Handler) Get(ctx echo.Context) error {
 		return webUtils.WriteErrorEchoServer(ctx, errors.New(constants.IndexOutOfRange), http.StatusBadRequest)
 	}
 
-	album, err := h.AlbumUseCase.GetById(id)
+	album, err := h.AlbumUseCase.GetById(int64(id))
 
 	if err != nil {
 		return webUtils.WriteErrorEchoServer(ctx, err, http.StatusBadRequest)
@@ -173,7 +173,7 @@ func (h Handler) Delete(ctx echo.Context) error {
 		return webUtils.WriteErrorEchoServer(ctx, errors.New(constants.IndexOutOfRange), http.StatusBadRequest)
 	}
 
-	if err := h.AlbumUseCase.Delete(id); err != nil {
+	if err := h.AlbumUseCase.Delete(int64(id)); err != nil {
 		return webUtils.WriteErrorEchoServer(ctx, err, http.StatusBadRequest)
 	}
