+++ internal/app/domain/album.go
@@ -7,29 +7,29 @@ import (
 )
 
 type Album struct {
-	Id             int    `json:"id" example:"8" db:"id" validate:"min=0,nonnil"`
+	Id             int64  `json:"id" example:"8" db:"id" validate:"min=0,nonnil"`
 	Title          string `json:"title" example:"Mercury" db:"title" validate:"max=256,nonnil"`
-	ArtistId       int    `json:"artistId" example:"4" db:"artist_id" validate:"min=0,nonnil"`
-	CountLikes     int    `json:"countLikes" example:"54" db:"count_likes" validate:"min=0,nonnil"`
-	CountListening int    `json:"countListening" example:"15632" db:"count_listening" validate:"min=0,nonnil"`
+	ArtistId       int64  `json:"artistId" example:"4" db:"artist_id" validate:"min=0,nonnil"`
+	CountLikes     int64  `json:"countLikes" example:"54" db:"count_likes" validate:"min=0,nonnil"`
+	CountListening int64  `json:"countListening" example:"15632" db:"count_listening" validate:"min=0,nonnil"`
 	Date           int64  `json:"date" example:"0" db:"date,nonnil"`
 }
 
 type AlbumRepo interface {
 	Insert(Album) error
 	Update(Album) error
-	Delete(int) error
-	SelectByID(int) (*Album, error)
+	Delete(int64) error
+	SelectByID(int64) (*Album, error)
 	GetAll() ([]Album, error)
 	GetPopular() ([]Album, error)
-	GetLastId() (id int, err error)
+	GetLastId() (id int64, err error)
 	//GetType() reflect.Type
-	GetSize() (int, error)
-	GetAlbumsFromArtist(artist int) ([]Album, error)
+	GetSize() (int64, error)
+	GetAlbumsFromArtist(artist int64) ([]Album, error)
 }
 
 type AlbumDataTransfer struct {
-	Id     int                 `json:"id" example:"1"`
+	Id     int64               `json:"id" example:"1"`
 	Title  string              `json:"title" example:"Mercury"`
 	Artist string              `json:"artist" example:"Hexed"`
 	Cover  string              `json:"cover" example:"assets/album_1.png"`
@@ -52,11 +52,11 @@ func (album *Album) CastToDtoWithoutArtistNameAndTracks() (*AlbumDataTransfer, e
 
 }
 
-func (album *Album) GetId() int {
+func (album *Album) GetId() int64 {
 	return album.Id
 }
 
-func (album *Album) SetId(id int) error {
+func (album *Album) SetId(id int64) error {
 	album.Id = id
 	return nil
 }
@@ -65,7 +65,7 @@ func (album *Album) Check() error {
 	return validator.Validate(album)
 }
 
-func (album *Album) GetCountListening() int {
+func (album *Album) GetCountListening() int64 {
 	return album.CountListening
 }
