+++ internal/app/domain/albumCover.go
@@ -7,16 +7,16 @@ import (
 type AlbumCoverRepo interface {
 	Insert(AlbumCover) error
 	Update(AlbumCover) error
-	Delete(int) error
-	SelectByID(int) (*AlbumCover, error)
+	Delete(int64) error
+	SelectByID(int64) (*AlbumCover, error)
 	GetAll() ([]AlbumCover, error)
-	GetLastId() (id int, err error)
+	GetLastId() (id int64, err error)
 	//GetType() reflect.Type
-	GetSize() (int, error)
+	GetSize() (int64, error)
 }
 
 type AlbumCover struct {
-	Id int `json:"id" example:"1" db:"id" validate:"min=0,nonnil"`
+	Id int64 `json:"id" example:"1" db:"id" validate:"min=0,nonnil"`
 	//Title  string `json:"title" example:"Mercury" db:"title" validate:"max=256,nonnil"`
 	Quote  string `json:"quote" example:"some phrases" db:"quote" validate:"max=512,nonnil"`
 	IsDark bool   `json:"isDark" example:"true" db:"is_dark" validate:"nonnil"`
@@ -28,11 +28,11 @@ type AlbumCoverDataTransfer struct {
 	IsDark bool   `json:"isDark" example:"true"`
 }
 
-func (cover *AlbumCover) GetId() int {
+func (cover *AlbumCover) GetId() int64 {
 	return cover.Id
 }
 
-func (cover *AlbumCover) SetId(id int) error {
+func (cover *AlbumCover) SetId(id int64) error {
 	cover.Id = id
 	return nil
 }
